Move CodeMirror LRLanguage to new file
This separates the base language support from the LSP and color picker.
