+++ src/editor/plugins/lsp/kcl/language.ts
@@ -3,11 +3,6 @@
 import {
   LRLanguage,
   LanguageSupport,
-  indentNodeProp,
-  continuedIndent,
-  delimitedIndent,
-  foldNodeProp,
-  foldInside,
 } from '@codemirror/language'
 import {
   LanguageServerClient,
@@ -15,9 +10,8 @@ import {
 } from '@kittycad/codemirror-lsp-client'
 import { kclPlugin } from '.'
 import type * as LSP from 'vscode-languageserver-protocol'
-// @ts-ignore: No types available
-import { parser } from './kcl.grammar'
 import { colorPicker } from './colors'
+import { KclLanguage } from './kcl'
 
 export interface LanguageOptions {
   workspaceFolders: LSP.WorkspaceFolder[]
@@ -28,31 +22,6 @@ export interface LanguageOptions {
     notification: LSP.NotificationMessage
   ) => void
 }
-
-export const KclLanguage = LRLanguage.define({
-  name: 'kcl',
-  parser: parser.configure({
-    props: [
-      indentNodeProp.add({
-        Body: delimitedIndent({ closing: '}' }),
-        BlockComment: () => null,
-        'Statement Property': continuedIndent({ except: /^{/ }),
-      }),
-      foldNodeProp.add({
-        'Body ArrayExpression ObjectExpression': foldInside,
-        BlockComment(tree) {
-          return { from: tree.from + 2, to: tree.to - 2 }
-        },
-        PipeExpression(tree) {
-          return { from: tree.firstChild!.to, to: tree.to }
-        },
-      }),
-    ],
-  }),
-  languageData: {
-    commentTokens: { line: '//', block: { open: '/*', close: '*/' } },
-  },
-})
 
 export function kcl(options: LanguageOptions) {
   return new LanguageSupport(KclLanguage, [
